codeit/미션18/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
openai
python-dotenv
